Move array tests to separate class
