+++ SQL-compiler/src/test/java/org/dbsp/sqlCompiler/compiler/ComplexQueriesTest.java
@@ -24,15 +24,12 @@
 package org.dbsp.sqlCompiler.compiler;
 
 import org.dbsp.sqlCompiler.circuit.DBSPCircuit;
-import org.dbsp.sqlCompiler.compiler.frontend.CalciteToDBSPCompiler;
 import org.dbsp.sqlCompiler.compiler.visitors.CircuitFunctionRewriter;
 import org.dbsp.sqlCompiler.compiler.visitors.DBSPCompiler;
 import org.dbsp.sqlCompiler.compiler.visitors.ThreeOperandVisitor;
 import org.dbsp.sqlCompiler.ir.expression.DBSPTupleExpression;
 import org.dbsp.sqlCompiler.ir.expression.literal.*;
 import org.dbsp.sqlCompiler.ir.type.primitive.DBSPTypeDouble;
-import org.dbsp.util.Logger;
-import org.junit.Assert;
 import org.junit.Test;
 
 @SuppressWarnings("SpellCheckingInspection")
@@ -63,54 +60,6 @@ public class ComplexQueriesTest extends BaseSQLTests {
         compiler.compileStatement(ddl);
         compiler.compileStatement(query);
         this.addRustTestCase(getCircuit(compiler));
-    }
-
-    @Test
-    public void testArray() {
-        String ddl = "CREATE TABLE ARR_TABLE (\n"
-                + "ID INTEGER,\n"
-                + "VALS INTEGER ARRAY,\n"
-                + "VALVALS VARCHAR(10) ARRAY)";
-        DBSPCompiler compiler = testCompiler();
-        compiler.compileStatement(ddl);
-        String query = "CREATE VIEW V AS SELECT *, CARDINALITY(VALS), ARRAY[ID, 5], VALS[1] FROM ARR_TABLE";
-        compiler.compileStatements(query);
-        if (compiler.hasErrors())
-            compiler.showErrors(System.err);
-        DBSPCircuit circuit = getCircuit(compiler);
-        this.addRustTestCase(circuit);
-    }
-
-    // @Test
-    public void testUnnest() {
-        // TODO: not yet implemented.
-        String ddl = "CREATE TABLE ARR_TABLE (\n"
-                + "ID INTEGER,\n"
-                + "VALS INTEGER ARRAY)";
-        DBSPCompiler compiler = testCompiler();
-        compiler.compileStatement(ddl);
-        String query = "CREATE VIEW V AS SELECT ID, VAL FROM ARR_TABLE, UNNEST(VALS) AS VAL";
-        Logger.INSTANCE.setDebugLevel(CalciteToDBSPCompiler.class, 3);
-        compiler.compileStatements(query);
-        if (compiler.hasErrors())
-            compiler.showErrors(System.err);
-        DBSPCircuit circuit = getCircuit(compiler);
-        this.addRustTestCase(circuit);
-    }
-
-    //@Test
-    public void test2DArray() {
-        // TODO: not yet implemented
-        String ddl = "CREATE TABLE ARR_TABLE (\n"
-                + "VALS INTEGER ARRAY ARRAY)";
-        DBSPCompiler compiler = testCompiler();
-        compiler.compileStatement(ddl);
-        String query = "CREATE VIEW V AS SELECT *, CARDINALITY(VALS), VALS[1] FROM ARR_TABLE";
-        compiler.compileStatements(query);
-        if (compiler.hasErrors())
-            compiler.showErrors(System.err);
-        DBSPCircuit circuit = getCircuit(compiler);
-        this.addRustTestCase(circuit);
     }
 
     @Test
