+++ SQL-compiler/src/test/java/org/dbsp/sqlCompiler/compiler/EndToEndTests.java
@@ -76,8 +76,10 @@ public class EndToEndTests extends BaseSQLTests {
     }
 
     //@Test
+    // Calcite's semantics requires this to crash at runtime;
+    // we should change the semantics of ELEMENT
+    // to return NULL when the array has more than 1 element.
     public void testArrayElement() {
-        // This should return null, but in calcite the result is not nullable!
         String query = "SELECT ELEMENT(ARRAY [2, 3])";
         DBSPZSetLiteral result =
                 new DBSPZSetLiteral(new DBSPTupleExpression(DBSPLiteral.none(DBSPTypeInteger.NULLABLE_SIGNED_32)));
